Move equipment module into items package
