+++ items/__init__.py
@@ -1,1 +1,2 @@
 from .item_data import Item, ALL_ITEMS
+from .equipment import Equipment, ALL_EQUIPMENT, CRAFTING_RECIPES
